+++ packages/backend/api/src/index.ts
@@ -16,6 +16,7 @@ const PORT = process.env.BACKEND_PORT ?? 3000;
 
 app.use(express.json());
 app.use(cors());
+
 app.use(
   '/cdn',
   express.static(
