+++ packages/backend/api/src/models/follow-model.ts
@@ -22,6 +22,15 @@ async function getFollowingCount(userId: number): Promise<number> {
   return Number(result?.count ?? 0);
 }
 
+// MODELE A REMPLACER
+// const result = await db
+// .selectFrom('follow_up as f')
+// .innerJoin('user as u', 'u.id', 'f.follower_id')
+// .select((eb) => eb.fn.countAll<number>().as('count'))
+// .where('f.followee_id', '=', userId)
+// .executeTakeFirst();
+// return result?.count ?? 0
+
 export default {
   getFollowersCount,
   getFollowingCount,
