+++ packages/backend/api/src/subrouters/follows-router.ts
@@ -7,7 +7,7 @@ const followsRouter = express.Router();
 const userId = 1;
 
 // GET /follows/:id/followers
-followsRouter.get('/:id/followers', async (req, res) => {
+followsRouter.get('/${userId}/followers', async (req, res) => {
   try {
     const count = await followModel.getFollowersCount(userId);
     res.json({ count });
@@ -18,7 +18,7 @@ followsRouter.get('/:id/followers', async (req, res) => {
 });
 
 // GET /follows/:id/following
-followsRouter.get('/:id/following', async (req, res) => {
+followsRouter.get('/:{userId}/following', async (req, res) => {
   try {
     const count = await followModel.getFollowingCount(userId);
     res.json({ count });
@@ -26,6 +26,28 @@ followsRouter.get('/:id/following', async (req, res) => {
     console.error('Erreur lors de la récupération des followings :', error);
     res.status(400).send('Erreur serveur');
   }
+});
+
+postsRouter.get('/`${userId}`/followers', async (req, res) => {
+  const userId = parseInt(req.params.id, 10);
+  if (isNaN(userId)) {
+    res.status(400).send('Invalid user ID');
+    return;
+  }
+
+  const followersCount = await followModel.getFollowersCount(userId);
+  res.json({ count: followersCount });
+});
+
+postsRouter.get('/`${userId}`/following', async (req, res) => {
+  const userId = parseInt(req.params.id, 10);
+  if (isNaN(userId)) {
+    res.status(400).send('Invalid user ID');
+    return;
+  }
+
+  const followingCount = await followModel.getFollowingCount(userId);
+  res.json({ count: followingCount });
 });
 
 export default followsRouter;
