+++ packages/frontend/web/src/app.tsx
@@ -1,26 +1,11 @@
-import { Outlet, useLocation } from 'react-router-dom';
+import { Outlet } from 'react-router-dom';
 
 import NavBar from './components/navbar';
 
 export default function App() {
-  const location = useLocation();
-
-  const hideMobileNavbarRoutes = [
-    '/parameters',
-    '/account-settings',
-    '/edit-bio',
-    '/edit-mail',
-    '/edit-username',
-    '/edit-password',
-    '/delete-account',
-    '/profile-informations',
-  ];
-
-  const isMobileHidden = hideMobileNavbarRoutes.includes(location.pathname);
-
   return (
     <div className='flex h-dvh w-full'>
-      <NavBar isMobileHidden={isMobileHidden} />
+      <NavBar />
       <main className='h-full w-full'>
         <Outlet />
       </main>
