+++ packages/frontend/web/src/main.tsx
@@ -1,4 +1,4 @@
-// import React from 'react';
+import React from 'react';
 import ReactDOM from 'react-dom/client';
 import { RouterProvider } from 'react-router-dom';
 
@@ -9,8 +9,8 @@ const rootElement = document.querySelector('#root');
 
 if (rootElement) {
   ReactDOM.createRoot(rootElement).render(
-    // <React.StrictMode>
-    <RouterProvider router={router} />,
-    // </React.StrictMode>,
+    <React.StrictMode>
+      <RouterProvider router={router} />
+    </React.StrictMode>,
   );
 }
