+++ packages/frontend/web/src/pages/profile-page.tsx
@@ -1,0 +1,12 @@
+import Profile from '@/components/profile';
+
+export default function ProfilePage() {
+  return (
+    <section className='flex'>
+      <Profile />
+    </section>
+  );
+}
+{
+  /* //size-68 sm:size-40  */
+}
