--- packages/frontend/web/src/pages/profile.tsx
@@ -1,12 +1,0 @@
-import ProfileOwn from '@/components/profile-own';
-
-export default function Profile() {
-  return (
-    <section className='flex'>
-      <ProfileOwn />
-    </section>
-  );
-}
-{
-  /* //size-68 sm:size-40  */
-}
